refactor: extract code into its own functions
