+++ packages/utils/src/cmds/transform_translate/alternative_language.template
@@ -1,0 +1,15 @@
+msgid ""
+msgstr ""
+"Project-Id-Version: PACKAGE VERSION\\n"
+"Report-Msgid-Bugs-To: \\n"
+"POT-Creation-Date: ###CREATION_DATE###\\n"
+"PO-Revision-Date: ###REVISION_DATE###\\n"
+"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n"
+"Language-Team: LANGUAGE <LL@li.org>\\n"
+"Language: ###LANGUAGE###\\n"
+"MIME-Version: 1.0\\n"
+"Content-Type: text/plain; charset=UTF-8\\n"
+"Content-Transfer-Encoding: 8bit\\n"
+"X-Generator: Translate Toolkit 2.2.5\\n"
+"X-Pootle-Path: /###LANGUAGE###/###APP_NAME###/###LANGUAGE###.po\\n"
+"X-Pootle-Revision: 29879\n"
