+++ packages/utils/src/cmds/transform_translate/main_language.template
@@ -1,0 +1,9 @@
+msgid ""
+msgstr ""
+"Project-Id-Version: i18next-conv\\n"
+"MIME-Version: 1.0\\n"
+"Content-Type: text/plain; charset=utf-8\\n"
+"Content-Transfer-Encoding: 8bit\\n"
+"Plural-Forms: nplurals=2; plural=(n != 1)\\n"
+"POT-Creation-Date: ###CREATION_DATE###\\n"
+"PO-Revision-Date: ###REVISION_DATE###\\n"
